+++ src/lib/terrain/voxelmap/viewer/voxelmap-viewer-base.ts
@@ -2,7 +2,7 @@ import { logger } from '../../../helpers/logger';
 import { createMeshesStatistics } from '../../../helpers/meshes-statistics';
 import * as THREE from '../../../libs/three-usage';
 import { type VoxelsChunkSize } from '../i-voxelmap';
-import { ChunkId } from '../patch/chunk-id';
+import { ChunkId } from '../chunk/chunk-id';
 import { EVoxelsDisplayMode } from '../voxelsRenderable/voxels-material';
 import { type VoxelsRenderable } from '../voxelsRenderable/voxels-renderable';
 import { VoxelsRenderableFactoryBase } from '../voxelsRenderable/voxelsRenderableFactory/voxels-renderable-factory-base';
